requirements.txt
runpod~=1.8.1
Pillow>=10.2.0
onnxruntime-gpu
